fixed for from size_t type to int

+++ include/task3.h
@@ -3,9 +3,10 @@
 template<typename T, size_t size>
 void map(T(&arr)[size],T (*change)(T))
 {
-	for (size_t i = 0; i < size; i++)
+	for (int i = 0; i < size; i++)
 	{
 		arr[i] = change(arr[i]);
 	} 
+	
 	return ;
 }
